ML/manual_data_fetch/requirements.txt
fastapi
uvicorn
python-dotenv
motor
pydantic<2.0
passlib[bcrypt]
python-jose[cryptography]
jinja2
python-multipart